+++ Documentation/source/changelog.md
@@ -16,7 +16,9 @@ __Github: [Issues](https://github.com/Egodystonic/TinyFFR/milestone/5?closed=1) 
 ### Improvements
 
 * Replaced the ambiguous "Recommended" display property on the display discoverer with separate "HighestResolution" and "HighestRefreshRate" properties
-* Added `Triangularize`/`TriangularizeRectified` methods to `Angle` to help create triangle functions
+* Math:
+	* Added `Triangularize`/`TriangularizeRectified` methods to `Angle` to help create triangle functions
+	* Added `UnitSphericalCoordinate` to represent unit-length spherical coord (azimuthal + polar angle); can be converted to a `Direction` given two basis directions
 
 ### Bug Fixes